requirements.txt
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
